requirements.txt
Django==1.11.6
django-filter==1.0.4
djangorestframework==3.7.0
olefile==0.44
Pillow==4.3.0
pytz==2017.2
